// old/types.ts
export type Product = {
	id: string;
	name: string;
	slug?: string;
	description: string;
	collections: Collection[];
	facetValues: FacetValue[];
	featuredAsset: FeaturedAsset;
	assets: Asset[];
	variants: Variant[];
	optionGroups: ProductOptionGroup[];
};

type Breadcrumb = {
	id: string;
	name: string;
	slug: string;
};

export type Collection = {
	id: string;
	slug: string;
	name: string;
	breadcrumbs?: Breadcrumb[];
	parent?: { name: '__root_collection__' };
	featuredAsset?: { id: string; preview: string };
	children: any[];
};

type Facet = {
	id: string;
	code: string;
	name: string;
};

type FacetValue = {
	facet: Facet;
	id: string;
	code: string;
	name: string;
};

type FeaturedAsset = {
	id: string;
	preview: string;
};

type Asset = {
	id: string;
	preview: string;
};

export type ProductOptionGroup = {
	id: string;
	code: string;
	name: string;
	options: ProductOption[];
};

export type ProductOption = {
	id: string;
	code: string;
	name: string;
	group?: ProductOptionGroup;
};

export type Variant = {
	id: string;
	name: string;
	priceWithTax: number;
	currencyCode: CurrencyCode;
	sku: string;
	stockLevel: string;
	trackInventory?: string | boolean;
	featuredAsset?: any;
	options: ProductOption[];
};